requirements.txt
flask
google-generativeai
nltk
transformers
torch
scikit-learn
pymongo
python-dotenv
fastapi
uvicorn[standard]
python-multipart
tensorflow